Extract FungibleItemValue
